ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.7.0
pymupdf>=1.25.0
python-dotenv>=1.0.1
llama-index-core>=0.12.0
